source/requirements.txt
aiofiles
aiohttp
asyncio
dnspython
pymongo[srv]
motor==3.0.0
psutil==5.9.1
TgCrypto
pykeyboard==0.1.5
python-arq==6.0.5
requests==2.27.1
python-dotenv==0.20.0
regex!=2021.8.27
cachetools==5.1.0
uvloop==0.16.0
pytz==2022.1
dateparser==1.1.1
gpytranslate==1.4.0
EmojiCaptcha==0.1.2
captcha==0.4 
googletrans==3.1.0a0
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor
pillow
psutil
py-tgcalls==0.8.6
pykeyboard
pyrogram==2.0.60
python-dotenv
pyyaml==5.3.1
requests
telegraph
telethon==1.25.2
gpytranslate
speedtest-cli
spotipy
tgcrypto
youtube-search
youtube-search-python
yt-dlp==2023.01.06
pymongo
apscheduler
telegraph[aio]
asyncstdlib
enums